Threaded Enemy manager and classes
Added an threaded enemy manager as well as an abstract enemy class and a
Normal enemy class. Nothing is truly connected to the main run loop yet
so this cannot break anything. All of this can be changed, but until we
get the world up and running I'd like to present this prototype.